+++ include/vxLib/Container/DynamicArray.h
@@ -39,7 +39,7 @@ namespace vx
 
 		VX_TYPE_INFO;
 
-		u64 m_blockSize;
+		size_t m_blockSize;
 		MyAllocator m_allocator;
 #ifdef _VX_ARRAY_ANALYZER
 		ArrayStats m_arrayStats;
@@ -164,7 +164,7 @@ namespace vx
 			}
 		}
 
-		void resize(u64 sz)
+		void resize(size_t sz)
 		{
 			auto currentSize = size();
 			if (currentSize >= sz)
@@ -180,7 +180,7 @@ namespace vx
 			}
 		}
 
-		void reserve(u64 c)
+		void reserve(size_t c)
 		{
 			auto currentCapacity = capacity();
 			if (c <= currentCapacity)
